changed type long to int

+++ src/main/java/ru/mudan/controllers/CalculateController.java
@@ -20,7 +20,7 @@ public class CalculateController {
     }
 
     @GetMapping
-    public ResponseEntity<VacationPayResponse>getVacationPay(@RequestParam Long averageSalary,
+    public ResponseEntity<VacationPayResponse>getVacationPay(@RequestParam Integer averageSalary,
                                                 @RequestParam Integer amountVacationDays){
         return new ResponseEntity<>(new VacationPayResponse(calculatorService.calculateVacationPay(averageSalary,amountVacationDays)), HttpStatus.OK);
     }
