+++ src/main/java/ru/mudan/response/VacationPayResponse.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class VacationPayResponse {
-    Long vacationPay;
+    Integer vacationPay;
 }
